requirements.txt
# Core Deep Learning
torch==2.0.1
torchvision==0.15.2
pytorch-lightning==2.0.9
torchmetrics==1.2.0

# Data Processing & Augmentation
numpy==1.24.4
Pillow==10.0.1
opencv-python==4.8.1.78
imgaug==0.4.0
albumentations==1.3.1
scipy==1.11.4

# Experiment Tracking & Hyperparameter Tuning
wandb==0.15.12
optuna==3.4.0

# Configuration Management
pyyaml==6.0.1
omegaconf==2.3.0

# Utilities
tqdm==4.66.1
matplotlib==3.7.3
seaborn==0.13.0
pandas==2.0.3

# Text Processing
nltk==3.8.1
wikipedia==1.4.0

# Development Tools
black==23.11.0
flake8==6.1.0
pytest==7.4.3

# Model Architectures (SwinTransformer)
timm==0.9.12

# Additional Utilities
scikit-learn==1.3.2
scikit-image==0.21.0
einops==0.7.0

# Additional dependencies for specific functionality
typing-extensions==4.8.0
setuptools==65.5.1